requirements.txt
# Elasticsearch
elasticsearch>=8.0.0

# Environment and configuration
python-dotenv>=1.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# Data generation
faker>=22.0.0

# Date handling
python-dateutil>=2.8.0

# HTTP requests (for Elastic Agent Builder API)
httpx>=0.25.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0

# Type hints
typing-extensions>=4.0.0
